+++ cmd/homeserver_offline_exporters/internal/archive_reader.go
@@ -1,4 +1,4 @@
-package _common
+package internal
 
 import (
 	"io"
