+++ cmd/homeserver_offline_exporters/internal/init_export.go
@@ -1,4 +1,4 @@
-package _common
+package internal
 
 import (
 	"flag"
